Backend/requirements.txt
fastapi
uvicorn
yt-dlp[default]
youtube-transcript-api
groq
python-dotenv
pydantic